requirements.txt
Flask==2.2.3
matplotlib==3.6.3
numpy==1.23.5
Pillow==9.4.0
Pillow==9.5.0
scipy==1.10.1
skimage==0.0
tensorflow==2.12.0rc0
tensorflow_intel==2.12.0rc0
Werkzeug==2.2.3
Jinja2==2.10.1
gunicorn==20.0.4
itsdangerous==1.1.0
MarkupSafe==1.1.1